Move debug output rendering to a dedicated HTML class still looks like crap...
